% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{asym_richness}
\alias{asym_richness}
\title{Asymptotic species richness estimates for the species discovery model.}
\usage{
asym_richness(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Asymptotic species richness estimates for the species discovery model.
}
