% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{asymptotic_richness}
\alias{asymptotic_richness}
\title{Extract the species richness estimates of the sdm output}
\usage{
asymptotic_richness(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Extract the species richness estimates of the sdm output
}
