% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPfit.R
\name{BNPfit}
\alias{BNPfit}
\title{Find the Empirical Bayes estimates of Bayesian Nonparametric Species Sampling model}
\usage{
BNPfit(species_counts, model, n_resamples = 500L)
}
\arguments{
\item{species_counts}{Vector of integer counts of the observed clusters (species)}

\item{model}{Model to fit. Available models are "Dirichlet", "Pitman-Yor", "LL2" and "LL3".}

\item{n_resamples}{Number of resamples of the original sequence. Needed for Non-exchangeable models only. Default = 500}
}
\value{
a list (?)
}
\description{
Find the Empirical Bayes estimates of Bayesian Nonparametric Species Sampling model
}
