% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{coef.sdm}
\alias{coef.sdm}
\title{Extract the coefficients of the sdm output}
\usage{
\method{coef}{sdm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Extract the coefficients of the sdm output
}
