% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage.DP}
\alias{coverage.DP}
\title{Coverage estimator for a Dirichlet process}
\usage{
\method{coverage}{DP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, DP}}

\item{...}{Additional parameters}
}
\description{
Coverage estimator for a Dirichlet process
}
\details{
Compute the posterior coverage estimator for the Dirichlet process.
}
