% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage.numeric}
\alias{coverage.numeric}
\title{Sample-based coverage estimator}
\usage{
\method{coverage}{numeric}(object, ...)
}
\description{
Sample-based coverage estimator
}
\details{
Compute the Touring estimator (Good, 1953) for the sample coverage.
}
