% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage.PY}
\alias{coverage.PY}
\title{Coverage estimator for a Pitman-Yor process}
\usage{
\method{coverage}{PY}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, PY}}

\item{...}{Additional parameters}
}
\description{
Coverage estimator for a Pitman-Yor process
}
\details{
Compute the posterior coverage estimator for the Pitman-Yor process.
}
