% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Sample-based or model-based coverage index}
\usage{
coverage(object, ...)
}
\arguments{
\item{object}{An object of class \code{numeric}, \code{ssm} or \code{sdm}}

\item{...}{Additional parameters}
}
\description{
Sample-based or model-based coverage index
}
