% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversities}
\alias{diversities}
\title{Diversity measures}
\usage{
diversities(dataset)
}
\arguments{
\item{frequencies}{A \code{K}-dimensional vector of frequencies
#'}
}
\value{
Collection of diversity indices
}
\description{
Diversity measures
}
