% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_discoveries.R
\name{extract_discoveries}
\alias{extract_discoveries}
\title{Obtain the sequence of new discoveries form}
\usage{
extract_discoveries(sequence)
}
\arguments{
\item{sequence}{A sequence of species in order of appearence}
}
\value{
A sequence of binary variables
}
\description{
Obtain the sequence of new discoveries form
}
\examples{
extract_discoveries(c("sp1", "sp1", "sp3", "sp4"))
}
