% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_miscellanea.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Build a random sequence of observed species form a vector of counts}
\usage{
extract_sequence(species_counts, seed = 42)
}
\arguments{
\item{species_counts}{Vector of counts of the species observed}

\item{seed}{Optional seed. Default is 42}
}
\value{
vector
}
\description{
Build a random sequence of observed species form a vector of counts
}
\examples{
build_random_sequences(c(1, 2, 45, 7))
}
