% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation}
\alias{extrapolation}
\title{Model base extrapolation curve and prediction}
\usage{
extrapolation(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm} or \code{sdm}}

\item{...}{Additional parameters}
}
\description{
Model base extrapolation curve and prediction
}
