% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{extrapolation.sdm}
\alias{extrapolation.sdm}
\title{Extrapolation function for a species discovery model.}
\usage{
\method{extrapolation}{sdm}(object, m, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{m}{Additional number of samples to predict.}

\item{...}{}
}
\description{
Extrapolation function for a species discovery model.
}
