% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LL2.R
\name{fit_LL2}
\alias{fit_LL2}
\title{Fit the two parameter log-logistic as in Zito et al. (2020)}
\usage{
fit_LL2(discoveries)
}
\arguments{
\item{discoveries}{Sequence of discoveries indicators}
}
\value{
Parameters of the log logistic distribution
}
\description{
Fit the two parameter log-logistic as in Zito et al. (2020)
}
\examples{
fit_LL2(c(1,0,1, 1,0,1, 0, 0,0))
}
