% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPfit_internal.R
\name{fit_LL3}
\alias{fit_LL3}
\title{Fit the three parameter log-logistic as in Zito et al. (2020)}
\usage{
fit_LL3(discoveries)
}
\arguments{
\item{discoveries}{Sequence of discovery indicators}
}
\value{
A list, including the parameters
}
\description{
Fit the three parameter log-logistic as in Zito et al. (2020)
}
\examples{
fit_LL3(c(1, 0, 1, 1, 0, 1, 0, 0, 0))
}
