% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_abundance.R
\name{freq_abundance.DP}
\alias{freq_abundance.DP}
\title{Model-based expected frequency of abundances for the Dirichlet process}
\usage{
\method{freq_abundance}{DP}(object, r, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, DP}}

\item{r}{frequencies of abundances}

\item{...}{Additional parameters}
}
\description{
Model-based expected frequency of abundances for the Dirichlet process
}
\details{
Specify the model based expected frequency of abundances for a
Dirichlet process
}
\examples{
r<- c(1:10)
fit <- ssm(fungalOTU, model = "DP")
freq_abundance(fit, r)
}
