% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_abundance.R
\name{freq_abundance.numeric}
\alias{freq_abundance.numeric}
\title{Frequency of abundances for a vector of species counts}
\usage{
\method{freq_abundance}{numeric}(object, r, ...)
}
\arguments{
\item{object}{An object of class \code{numeric} containing species abundances}

\item{r}{frequencies of abundances}

\item{...}{Additional parameters}
}
\description{
Frequency of abundances for a vector of species counts
}
\details{
Return the table of the species counts
}
