% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_abundance.R
\name{freq_abundance.PY}
\alias{freq_abundance.PY}
\title{Model-based expected frequency of abundances for the Pirman-Yor process}
\usage{
\method{freq_abundance}{PY}(object, r, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, PY}}

\item{r}{frequencies of abundances}

\item{...}{Additional parameters}
}
\description{
Model-based expected frequency of abundances for the Pirman-Yor process
}
\details{
Specify the model based expected frequency of abundances for a
Pitman-Yor process
}
\examples{
r<- c(1:10)
fit <- ssm(fungalOTU, model = "PY")
freq_abundance(fit, r)
}
