% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_abundance.R
\name{freq_abundance}
\alias{freq_abundance}
\title{Frequency of abundances}
\usage{
freq_abundance(object, ...)
}
\arguments{
\item{object}{An object of class \code{numeric} or \code{ssm}}

\item{...}{Additional parameters}
}
\description{
Frequency of abundances
}
