% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{freq_of_freq}
\alias{freq_of_freq}
\title{Diversity measures}
\usage{
freq_of_freq(dataset, rel = FALSE, plot = FALSE)
}
\arguments{
\item{rel}{a}

\item{plot}{a}

\item{frequencies}{A \code{K}-dimensional vector of frequencies}
}
\value{
Collection of diversity indices
}
\description{
Diversity measures
}
