% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fungalOTU}
\alias{fungalOTU}
\title{Abundances of distinct fungal OTUs}
\format{
A numeric vector
}
\usage{
fungalOTU
}
\description{
A dataset of abundances of fungal species from a field experiment.
The total abundance is 21,243, divided between 563 distinct species
}
\keyword{datasets}
