% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini.DP}
\alias{Gini.DP}
\title{Posterior Gini heterogeneity index for the Dirichlet process.}
\usage{
\method{Gini}{DP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, DP}.}

\item{...}{Additional parameters}
}
\description{
Posterior Gini heterogeneity index for the Dirichlet process.
}
