% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini.numeric}
\alias{Gini.numeric}
\title{Gini heterogeneity index for a vector pf species abundances}
\usage{
\method{Gini}{numeric}(object, ...)
}
\arguments{
\item{object}{An object of class \code{numeric}}

\item{...}{Additional parameters}
}
\description{
Gini heterogeneity index for a vector pf species abundances
}
