% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini.PY}
\alias{Gini.PY}
\title{Posterior Gini heterogeneity index for the Pitman-Yor process}
\usage{
\method{Gini}{PY}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, PY}}

\item{...}{Additional parameters}
}
\description{
Posterior Gini heterogeneity index for the Pitman-Yor process
}
