% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini}
\alias{Gini}
\title{Sample-based or model-based Gini heterogeniety index}
\usage{
Gini(object, ...)
}
\arguments{
\item{object}{An object of class \code{numeric}, \code{ssm} or \code{sdm}}

\item{...}{Additional parameters}
}
\description{
Sample-based or model-based Gini heterogeniety index
}
