% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lepidoptera}
\alias{Lepidoptera}
\title{Abundances of the Malayse butterfly dataset of Fisher et al. (1943)}
\format{
A numeric vector
}
\usage{
Lepidoptera
}
\description{
A dataset of abundances of Lepidoptera species derived from Fisher et al. (1943)
}
\keyword{datasets}
