% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{logLik.sdm}
\alias{logLik.sdm}
\title{Extract the loglikelihood of the sdm output}
\usage{
\method{logLik}{sdm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Extract the loglikelihood of the sdm output
}
