% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{plot.sdm}
\alias{plot.sdm}
\title{Compute rarefaction curves}
\usage{
\method{plot}{sdm}(object, n_points = 100, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{n_points}{Number of points to plot in the accumulation curve}

\item{...}{other parameters}
}
\description{
Compute rarefaction curves
}
