% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{predict.DP}
\alias{predict.DP}
\title{Predict method for the DP model}
\usage{
\method{predict}{DP}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ssm]{ssm}}.}

\item{newdata}{A new data frame containing}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predict method for a DP model, producing either the rarefaction curve or the extrapolation curve.
}
\details{
The method...
}
