% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{predict.sdm}
\alias{predict.sdm}
\title{Predict sdm}
\usage{
\method{predict}{sdm}(object, ...)
}
\arguments{
\item{object}{object of class \code{\link[sdm]{sdm}}}

\item{...}{additional values}
}
\description{
Predict sdm
}
