% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{print.summary.sdm}
\alias{print.summary.sdm}
\title{Print method for the summary}
\usage{
\method{print}{summary.sdm}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[summary.sdm]{summary.sdm}}}

\item{...}{other parameters}
}
\description{
Print method for the summary
}
