% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.numeric}
\alias{rarefaction.numeric}
\title{Sample-based rarefaction curve for a vector of species counts}
\usage{
\method{rarefaction}{numeric}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An vector of frequencies}

\item{verbose}{If TRUE, the estimation process will be printed}

\item{...}{Additional parameters}
}
\description{
Sample-based rarefaction curve for a vector of species counts
}
