% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.PY}
\alias{rarefaction.PY}
\title{Model-based rarefaction curve for a Dirichlet process model}
\usage{
\method{rarefaction}{PY}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{ssm}, \link[sdm]{PY}}}

\item{...}{Additional parameters}
}
\description{
Model-based rarefaction curve for a Dirichlet process model
}
