% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction}
\alias{rarefaction}
\title{Sample-based or model-based rarefaction curve}
\usage{
rarefaction(object, ...)
}
\arguments{
\item{object}{An object of class \code{numeric}, \code{ssm} or \code{sdm}}

\item{...}{Additional parameters}
}
\description{
Sample-based or model-based rarefaction curve
}
