% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.sdm}
\alias{rarefaction.sdm}
\title{Model-based rarefaction curve for a Sequential Discovery Model}
\usage{
\method{rarefaction}{sdm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{...}{Additional parameters}
}
\description{
Model-based rarefaction curve for a Sequential Discovery Model
}
