% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{rcoverage.DP}
\alias{rcoverage.DP}
\title{Random samples from the coverage posterior distribution for a Dirichlet process}
\usage{
\method{rcoverage}{DP}(object, n_samples = 1000, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, DP}}

\item{n_samples}{Number of samples to draw}

\item{...}{Additional parameters}
}
\description{
Random samples from the coverage posterior distribution for a Dirichlet process
}
