% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{rcoverage.PY}
\alias{rcoverage.PY}
\title{Random samples from the coverage posterior distribution for a Pitman-Yor process}
\usage{
\method{rcoverage}{PY}(object, n_samples = 1000, ...)
}
\arguments{
\item{object}{An object of class \code{ssm, PY}}

\item{n_samples}{Number of samples to draw}

\item{...}{Additional parameters}
}
\description{
Random samples from the coverage posterior distribution for a Pitman-Yor process
}
