% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{rcoverage}
\alias{rcoverage}
\title{Random sampler for the posterior distribution of a model-based coverage}
\usage{
rcoverage(object, ...)
}
\arguments{
\item{object}{An object of class \code{sdm}}

\item{...}{Additional parameters}
}
\description{
Random sampler for the posterior distribution of a model-based coverage
}
