% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random sampler from ad Dirichlet distribution}
\usage{
rdirichlet(size, alpha)
}
\arguments{
\item{size}{Number of samples to draw}

\item{alpha}{Centrality parameter. Has to be a vector}
}
\value{
Samples from a dirichlet distribution
}
\description{
Random sampler from ad Dirichlet distribution
}
