% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPY.R
\name{rDP}
\alias{rDP}
\title{Random samples from a Pitman-Yor process}
\usage{
rDP(size, alpha)
}
\arguments{
\item{size}{Number of observations to draw}

\item{alpha}{Centrality parameter}
}
\value{
A vector of frequencies
}
\description{
Random samples from a Pitman-Yor process
}
