% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{rGini.DP}
\alias{rGini.DP}
\title{Random sampler for the posterior distribution of the Gini heterogeneity index under a Dirichlet process}
\usage{
\method{rGini}{DP}(
  object,
  n_samples = 100,
  truncation_method = "dynamic",
  prob_cutoff = 0.99,
  static_cutoff = 1000,
  ...
)
}
\arguments{
\item{object}{An object of class \code{ssm, DP}}

\item{n_samples}{Number of posterior samples to draw}

\item{truncation_method}{Truncation method for the stick breaking sampler}

\item{prob_cutoff}{Probabilistic truncation cutoff for the stick breaking sampler. Valid only for \code{truncation_method = "dynamic"}}

\item{static_cutoff}{Integer cutoff for the stick breaking sampler. Valid only for \code{truncation_method = "static"}}

\item{...}{Additional model parameters}
}
\description{
Random sampler for the posterior distribution of the Gini heterogeneity index under a Dirichlet process
}
