% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{rGini}
\alias{rGini}
\title{Random sampler for the posterior distribution of the Gini heterogeneity index}
\usage{
rGini(object, ...)
}
\arguments{
\item{object}{An object of class \code{ssm}}

\item{...}{Additional parameters}
}
\description{
Random sampler for the posterior distribution of the Gini heterogeneity index
}
