% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPY.R
\name{rPY}
\alias{rPY}
\title{Random samples from a Pitman-Yor process}
\usage{
rPY(size, alpha, sigma)
}
\arguments{
\item{size}{Number of observations to draw}

\item{alpha}{Centrality parameter}

\item{sigma}{Discount parameter}
}
\value{
A vector of frequencies
}
\description{
Random samples from a Pitman-Yor process
}
