% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStickBreaking.R
\name{rStickBreaking}
\alias{rStickBreaking}
\title{Random Sampler for a Stick breaking construction of a Pitman-Yor}
\usage{
rStickBreaking(
  alpha,
  sigma,
  truncation_method = "dynamic",
  prob_cutoff = 0.99,
  static_cutoff = 1000
)
}
\arguments{
\item{alpha}{Centrality parameter of the Pitman-Yor}

\item{sigma}{Tail parameter of the Pitman-Yor}

\item{truncation_method}{Howe to truncate the stick breaking.
Available options are \code{"dynamic"} and \code{"static"}.
The first one stops sampling the from the distribution when the sampled weights
sum is higher than the \code{prob_cutoff}. The second instead performs a
hard truncation at the specified \code{static_cutoff}.}

\item{prob_cutoff}{Probability cutoff to stop the sampler. Valid for \code{truncation_method = "dynamic"}}

\item{static_cutoff}{Integer cutoff to stop the sampler. Valid for \code{truncation_method = "static"}}
}
\value{
A vector of weights representig the Pitman-Yor weights. Notice that the last element in the vector represents the aggregate weight after the truncation.
}
\description{
Random Sampler for a Stick breaking construction of a Pitman-Yor
}
