% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Kinf.R
\name{sample_Kinf}
\alias{sample_Kinf}
\title{Sample from the posterior distribution of the asymptotic species richness in a species discovery model}
\usage{
sample_Kinf(object, n_samples = 100, tolerance = 1e-07, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}}

\item{n_samples}{Number of random samples to draw}

\item{tolerance}{Tolerance to determine the truncation point. Default \code{= 1e-7}.}

\item{...}{Additional model parameters}
}
\value{
A vector of size \code{n_samples}
}
\description{
Sample from the posterior distribution of the asymptotic species richness in a species discovery model
}
\details{
The function sample from the posterior distribution of the asymptotic
         species richness under a species discovery model. Specifyng the tolerance determines the truncation point.
}
