% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_miscellanea.R
\name{sample_sequence}
\alias{sample_sequence}
\title{Build a random sequence of observed species form a vector of counts}
\usage{
sample_sequence(frequencies)
}
\arguments{
\item{frequencies}{Vector of counts of the species observed}
}
\value{
vector
}
\description{
Build a random sequence of observed species form a vector of counts
}
