% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{sdm}
\alias{sdm}
\title{Sequential discovery model as in Zito et al. (2020+)}
\usage{
sdm(frequencies, model = "LL3", verbose = TRUE)
}
\arguments{
\item{frequencies}{Vector of frequencies of the observed species}

\item{model}{Model to fit. Options are "LL3" and "Weibull"}

\item{verbose}{if TRUE, monitor the construction of the average rarefaction curve.}
}
\description{
Sequential discovery model as in Zito et al. (2020+)
}
\details{
This function fits the sequential discovery model is Zito et al. (2020+) on the average rarefaction curve obtained from a vector of species frequencies.
         The models available are "LL3" and "Weibull".
}
\examples{
# Fit the model
frequencies <- fungalOTU
fit <- sdm(frequencies, model = "LL3")
summary(fit)

}
