% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{summary.PY}
\alias{summary.PY}
\title{Heterogeneity for the PY model}
\usage{
\method{summary}{PY}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ssm]{ssm}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Gini heterogeneity
}
\details{
The method...
}
