% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage.sdm}
\alias{coverage.sdm}
\title{Coverage estimator for a sequential discovery model}
\usage{
\method{coverage}{sdm}(object, ...)
}
\description{
Coverage estimator for a sequential discovery model
}
\details{
Compute the posterior coverage estimator for a sequential discovery model
}
