% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation.sdm}
\alias{extrapolation.sdm}
\title{Extrapolation function for a species discovery model.}
\usage{
\method{extrapolation}{sdm}(object, m, ...)
}
\arguments{
\item{object}{An object of class \code{sdm}.}

\item{m}{Additional number of samples to predict.}

\item{...}{Additional parameters}
}
\description{
Extrapolation function for a species discovery model.
}
