% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{plot.DP}
\alias{plot.DP}
\title{Plot for the DP model}
\usage{
\method{plot}{DP}(object, type = "rarefaction", ...)
}
\arguments{
\item{object}{An object of class \code{\link[ssm]{ssm}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot for the DP model
}
\details{
The method...
}
