% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{plot.sdm}
\alias{plot.sdm}
\title{Compute rarefaction curves}
\usage{
\method{plot}{sdm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{newdata}{a new dataset}

\item{...}{other parameters}
}
\description{
Compute rarefaction curves
}
