% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{plot.sdm}
\alias{plot.sdm}
\title{Plot the average rarefaction curve and the fitted one (in red)}
\usage{
\method{plot}{sdm}(x, n_points = 100, type = "rarefaction", m = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link[sdm]{sdm}}.}

\item{n_points}{Number of points to plot in the accumulation curve}

\item{type}{Type of curve to plot. Available options are "rarefaction" and "extrapolation".
In the second case, one needs to provide the additional number of samples m up to which extrapolate.}

\item{m}{Additional number of samples. Required for extrapolation only}

\item{...}{additional parameters}
}
\description{
Plot the average rarefaction curve and the fitted one (in red)
}
