% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.DP}
\alias{rarefaction.DP}
\title{Model-based rarefaction curve for a Dirichlet process model}
\usage{
\method{rarefaction}{DP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{ssm}, \link[sdm]{DP}}}
}
\description{
Model-based rarefaction curve for a Dirichlet process model
}
