% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.numeric}
\alias{rarefaction.numeric}
\title{Sample-based rarefaction curve for a vector of species counts}
\usage{
\method{rarefaction}{numeric}(frequencies, verbose = TRUE, ...)
}
\description{
Sample-based rarefaction curve for a vector of species counts
}
