% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{rarefaction.sdm}
\alias{rarefaction.sdm}
\title{Extract the predicted rarefaction curve from a species discovery model. This is equivalent to the prediction function with default parameters}
\usage{
\method{rarefaction}{sdm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{...}{Additional parameters}
}
\description{
Extract the predicted rarefaction curve from a species discovery model. This is equivalent to the prediction function with default parameters
}
