% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{saturation}
\alias{saturation}
\title{Saturation for a sequential discovery model}
\usage{
saturation(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Saturation for a sequential discovery model
}
