% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{sdm}
\alias{sdm}
\title{Fit the sequential discovery model of Zito et al. (2020+)}
\usage{
sdm(frequencies, n_resamples = 500L, model = "LL3", verbose = TRUE)
}
\arguments{
\item{frequencies}{Counts of the species}

\item{n_resamples}{number of accumulation curves to generate}

\item{model}{model to fit. Options are "LL3" and "Weibull"}

\item{verbose}{Whether to monitor the state of the simulation or not}
}
\description{
Fit the sequential discovery model of Zito et al. (2020+)
}
\details{
This function runs the sequential discovery model is Zito et al. (2020+)
}
