% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{sdm}
\alias{sdm}
\title{Fit the sequential discovery model of Zito et al. (2020+)}
\usage{
sdm(frequencies, n_resamples = 100L, verbose = TRUE)
}
\arguments{
\item{frequencies}{Counts of the species}

\item{n_resamples}{number of accumulation curves to generate}

\item{verbose}{Whether to monitor the function or not}
}
\description{
Fit the sequential discovery model of Zito et al. (2020+)
}
\details{
This function.
}
