% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{Species sampling model}
\usage{
ssm(frequencies, model)
}
\arguments{
\item{frequencies}{A \code{K}-dimensional vector of frequencies}

\item{model}{Model to fit. Available models are \code{DP} (Dirichlet Process) and \code{PY} (Pitman-Yor process)}
}
\value{
An object of class \code{ssm}
}
\description{
Species sampling model
}
