% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{Find the Empirical Bayes estimates of Bayesian Nonparametric Species Sampling model}
\usage{
ssm(frequencies, model)
}
\arguments{
\item{frequencies}{A \code{K}-dimensional vector of frequencies}

\item{model}{Model to fit. Available models are "DP" (Dirichlet Process) and "PY" (Pitman-Yor process)
#'}
}
\value{
An object of class "ssm"
}
\description{
Find the Empirical Bayes estimates of Bayesian Nonparametric Species Sampling model
}
