% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{summary.sdm}
\alias{summary.sdm}
\title{Summary for a species discovery model}
\usage{
\method{summary}{sdm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[sdm]{sdm}}.}

\item{...}{additional parameters}
}
\description{
Summary for a species discovery model
}
\details{
A function to summarize the three parameter log-logistic output
}
